separated login function inside helper folder
