+++ tests/aam/authentication/nurse.spec.js
@@ -1,0 +1,9 @@
+import { test, expect } from '@playwright/test';
+import { faker } from '@faker-js/faker';
+import { login } from '../../../helpers/login.js';
+
+
+test('Creating the beneficiary', async ({ page }) => {
+  await login(page);
+  await page.getByRole('button', { name: 'Continue' }).click();
+});
